# perf: move to filter map from map and filter