--- a/rust/matcher.rs
+++ b/rust/matcher.rs
@@ -17,8 +17,7 @@ impl Matcher {
 
     pub fn score(&self, text: &str) -> i64 {
         self.matcher
-            .fuzzy_indices(text, &self.pattern)
-            .map(|(score, _indices)| score)
+            .fuzzy_match(text, &self.pattern)
             .unwrap_or_default()
     }
 }